[Deploy] - dev -> main으로 merge

+++ src/main/java/com/heartz/byeboo/adapter/in/web/dto/response/StepResponseDto.java
@@ -6,11 +6,11 @@ import com.heartz.byeboo.domain.type.EStep;
 import java.util.List;
 
 public record StepResponseDto(
-        Integer stepNumber,
+        Long stepNumber,
         String step,
         List<AllQuestDetailResponseDto> quests
 ) {
-    public static StepResponseDto of(Integer stepNumber, EStep step, List<AllQuestDetailResponseDto> quests) {
+    public static StepResponseDto of(Long stepNumber, EStep step, List<AllQuestDetailResponseDto> quests) {
         return new StepResponseDto(stepNumber, step.getLabel(), quests);
     }
 }
